Visualizations and graph library
