+++ frontend/public/locales/de/backend.json5
@@ -1,9 +1,12 @@
 {
   'infection-states': {
     Infected: 'Infiziert',
+    MildInfections: 'Milde Infektion',
     Hospitalized: 'Hospitalisiert',
     ICU: 'Intensivpatient',
     Dead: 'Verstorben',
+    DeadV1: 'Verstorben (💉)',
+    DeadV2: 'Verstorben (💉💉)',
     Exposed: 'Kontaktperson',
     Recovered: 'Genesen',
     Carrier: 'Träger',
@@ -26,6 +29,13 @@
     CarrierV2: 'Träger (💉💉)',
     SusceptibleV1: 'Empfänglich (💉)',
     SusceptibleV2: 'Empfänglich (💉💉)',
+    tooltip: 'Die hier angezeigten Zahlen zu Infizierten und Hospitalisierten basieren auf Annahmen zur Immunität,\
+              Dauer und Schweregrad einer Infektion mit Sars-CoV-2. Sie berechnen sich über ein komplexes mehrstufiges\
+              Verfahren aus den berichteten Fallzahlen und eine einfach Formel kann nicht angegeben werden. Ein\
+              Anstieg/Rückgang der Fallzahlen bedeutet einen Anstieg/Rückgang der Anzahl Infizierter, sodass in\
+              Relation zu den Werten der vergangenen Tagen bewertet werden kann, ob die Situation sich verbessert oder\
+              verschlimmert. Sofern verfügbar, kommen die Zahlen zu den Intensivpatient:innen aus dem DIVI\
+              Intensivregister.',
   },
   'scenario-names': {
     'Summer 2021 Simulation 1': 'Szenario ohne Maßnahmen',
