+++ frontend/src/components/Sidebar/DistrictMap.tsx
@@ -18,6 +18,7 @@ import {HeatmapLegend} from '../../types/heatmapLegend';
 import LockOpen from '@mui/icons-material/LockOpen';
 import LoadingContainer from '../shared/LoadingContainer';
 import mapData from 'assets/lk_germany_reduced.geojson';
+import {useGetCaseDataByDateQuery} from '../../store/services/caseDataApi';
 
 const {useRef} = React;
 
@@ -46,24 +47,57 @@ export default function DistrictMap(): JSX.Element {
   const legend = useAppSelector((state) => state.userPreference.selectedHeatmap);
   const [chart, setChart] = useState<am5map.MapChart | null>(null);
 
-  const {data, isFetching} = useGetSimulationDataByDateQuery(
+  const {data: simulationData, isFetching: isSimulationDataFetching} = useGetSimulationDataByDateQuery(
     {
       id: selectedScenario ?? 0,
       day: selectedDate ?? '',
       groups: ['total'],
       compartments: [selectedCompartment ?? ''],
     },
-    {skip: !selectedScenario || !selectedCompartment || !selectedDate}
+    {skip: selectedScenario === null || selectedScenario === 0 || !selectedCompartment || !selectedDate}
+  );
+
+  const {data: caseData, isFetching: isCaseDataFetching} = useGetCaseDataByDateQuery(
+    {
+      day: selectedDate ?? '',
+      groups: ['total'],
+      compartments: [selectedCompartment ?? ''],
+    },
+    {skip: selectedScenario === null || selectedScenario > 0 || !selectedCompartment || !selectedDate}
   );
 
   const legendRef = useRef<am5.HeatLegend | null>(null);
   const {t, i18n} = useTranslation();
   const {t: tBackend} = useTranslation('backend');
-  const {formatNumber} = NumberFormatter(i18n.language, 3, 8);
+  const {formatNumber} = NumberFormatter(i18n.language, 1, 0);
   const theme = useTheme();
   const dispatch = useAppDispatch();
   const lastSelectedPolygon = useRef<am5map.MapPolygon | null>(null);
   const [fixedLegendMaxValue, setFixedLegendMaxValue] = useState<number | null>(null);
+
+  // This memo either returns the case data or simulation data, depending on which card is selected.
+  const data = useMemo(() => {
+    if (selectedScenario === null) {
+      return null;
+    }
+
+    if (selectedScenario === 0 && caseData !== undefined && caseData.results.length > 0) {
+      return caseData;
+    } else if (selectedScenario > 0 && simulationData !== undefined && simulationData.results.length > 0) {
+      return simulationData;
+    }
+
+    return null;
+  }, [caseData, simulationData, selectedScenario]);
+
+  // This memo returns if the required data is currently being fetched. Either the case data or the scenario data.
+  const isFetching = useMemo(() => {
+    if (selectedScenario === null) {
+      return true;
+    }
+
+    return (selectedScenario === 0 && isCaseDataFetching) || (selectedScenario > 0 && isSimulationDataFetching);
+  }, [selectedScenario, isCaseDataFetching, isSimulationDataFetching]);
 
   // use Memoized to store aggregated max and only recalculate if parameters change
   const aggregatedMax: number = useMemo(() => {
@@ -71,7 +105,8 @@ export default function DistrictMap(): JSX.Element {
       return fixedLegendMaxValue;
     }
     let max = 1;
-    if (data && selectedCompartment) {
+
+    if (data && selectedCompartment !== null) {
       data.results.forEach((entry) => {
         if (entry.name !== '00000') {
           max = Math.max(entry.compartments[selectedCompartment], max);
@@ -177,8 +212,8 @@ export default function DistrictMap(): JSX.Element {
     });
     setChart(chart);
     return () => {
-      chart && chart.dispose();
-      root && root.dispose();
+      chart?.dispose();
+      root?.dispose();
     };
   }, [geodata, theme, t, formatNumber, dispatch]);
 
@@ -218,7 +253,7 @@ export default function DistrictMap(): JSX.Element {
   useEffect(() => {
     if (chart && chart.series.length > 0) {
       const polygonSeries = chart.series.getIndex(0) as am5map.MapPolygonSeries;
-      if (selectedScenario && selectedCompartment && !isFetching) {
+      if (selectedScenario !== null && selectedCompartment && !isFetching && data) {
         // Map compartment value to RS
         const dataMapped = new Map<string, number>();
         data?.results.forEach((entry) => {
@@ -229,8 +264,7 @@ export default function DistrictMap(): JSX.Element {
         if (dataMapped.size > 0) {
           polygonSeries.mapPolygons.each((polygon) => {
             const regionData = polygon.dataItem?.dataContext as IRegionPolygon;
-            const temp = dataMapped.get(regionData.RS);
-            regionData.value = temp != undefined ? temp : Number.NaN;
+            regionData.value = dataMapped.get(regionData.RS) ?? Number.NaN;
 
             // determine fill color
             let fillColor = am5.color(theme.palette.background.default);
@@ -244,28 +278,27 @@ export default function DistrictMap(): JSX.Element {
               }
             }
 
+            const bez = t(`BEZ.${regionData.BEZ}`);
+            const compartmentName = tBackend(`infection-states.${selectedCompartment}`);
             polygon.setAll({
               tooltipText:
-                selectedScenario && selectedCompartment
-                  ? `${t(`BEZ.${regionData.BEZ}`)} {GEN}\n${tBackend(
-                      `infection-states.${selectedCompartment}`
-                    )}: ${formatNumber(regionData.value)}`
-                  : `${t(`BEZ.${regionData.BEZ}`)} {GEN}`,
+                selectedScenario !== null && selectedCompartment
+                  ? `${bez} {GEN}\n${compartmentName}: ${formatNumber(regionData.value)}`
+                  : `${bez} {GEN}`,
               fill: fillColor,
             });
           });
         }
-      } else {
-        if (longLoad) {
-          polygonSeries.mapPolygons.each((polygon) => {
-            const regionData = polygon.dataItem?.dataContext as IRegionPolygon;
-            regionData.value = Number.NaN;
-            polygon.setAll({
-              tooltipText: `${t(`BEZ.${regionData.BEZ}`)} {GEN}`,
-              fill: am5.color(theme.palette.text.disabled),
-            });
+      } else if (longLoad || !data) {
+        polygonSeries.mapPolygons.each((polygon) => {
+          const regionData = polygon.dataItem?.dataContext as IRegionPolygon;
+          regionData.value = Number.NaN;
+          const bez = t(`BEZ.${regionData.BEZ}`);
+          polygon.setAll({
+            tooltipText: `${bez} {GEN}`,
+            fill: am5.color(theme.palette.text.disabled),
           });
-        }
+        });
       }
     }
   }, [
