+++ frontend/src/store/ScenarioSlice.ts
@@ -31,7 +31,17 @@ export const ScenarioSlice = createSlice({
 export const {setScenarios, setCompartments} = ScenarioSlice.actions;
 export default ScenarioSlice.reducer;
 
-const preferredOrder = ['Infected', 'Hospitalized', 'ICU', 'Dead', 'Exposed', 'Recovered', 'Carrier', 'Susceptible'];
+const preferredOrder = [
+  'MildInfections',
+  'Infected',
+  'Hospitalized',
+  'ICU',
+  'Dead',
+  'Exposed',
+  'Recovered',
+  'Carrier',
+  'Susceptible',
+];
 
 /** This function sorts an array with the first elements being the above preference if they are included in the array. */
 function sortWithPreference(array: Array<string>): Array<string> {
