+++ frontend/src/util/Theme.ts
@@ -124,11 +124,20 @@ export default createTheme({
         },
       },
     },
+    MuiTooltip: {
+      styleOverrides: {
+        tooltip: {
+          fontSize: '10pt',
+          textAlign: 'justify',
+        },
+      },
+    },
   },
   custom: {
     //scenario colors with graduations
     scenarios: [
       // Main   | Grad-1   | Grad-2   | Grad-3   | Grad-4   | Grad-5   | Disabled  |
+      ['#000000', '#303030', '#606060', '#909090', '#c0c0c0', '#e0e0e0', '#8c8c8c'], // Case Data
       ['#3998db', '#56a8e3', '#6cb6eb', '#7fc1f0', '#94cef7', '#b8e1ff', '#8c8c8c'], // Scenario 1
       ['#876be3', '#987feb', '#a892f0', '#b5a2f5', '#c0affa', '#d0c2ff', '#8c8c8c'], // Scenario 2
       ['#cc5ac7', '#e070db', '#eb7ce6', '#f291ee', '#faa0f6', '#fcb6f9', '#8c8c8c'], // Scenario 3
